+++ src/Config/Option/UrlShortener/AutoResolveTitlesConfigOption.php
@@ -8,16 +8,19 @@ use Shlinkio\Shlink\Config\Collection\PathCollection;
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-/** @deprecated */
-class ValidateUrlConfigOption extends BaseConfigOption
+class AutoResolveTitlesConfigOption extends BaseConfigOption
 {
     public function getConfigPath(): array
     {
-        return ['url_shortener', 'validate_url'];
+        return ['url_shortener', 'auto_resolve_titles'];
     }
 
     public function ask(StyleInterface $io, PathCollection $currentOptions): bool
     {
-        return $io->confirm('Do you want to validate long urls by 200 HTTP status code on response?');
+        return $io->confirm(
+            'Do you want Shlink to resolve the short URL title based on the long URL \'s title tag (if any)? '
+                . 'Otherwise, it will be kept empty unless explicitly provided.',
+            false,
+        );
     }
 }